+++ CHANGELOG.md
@@ -9,6 +9,7 @@
 
 ### On Progress
 
+- Remove unused lib: Lodash
 - Investigate error: "Failed to find response element - 842"
 
 ### Planned